#255 Convert to hook
